Extract classes to lib
