+++ lib/index.rb
@@ -1,0 +1,5 @@
+class Index < DropboxFile
+  def initialize
+    @path = "/index.md"
+  end
+end
